pyproject.toml
[tool.poetry]
name = "simbirsoft_sdet"
version = "0.1.0"
description = ""
authors = ["Irek Gimatov <gimatov@list.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
python-dotenv = "^0.19.0"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
